+++ lib/WindowsCPP.cpp
@@ -31,44 +31,44 @@ extern "C" {
 
 #endif
 
-const wstring _achartowstring(const char *s, int len) {
+const wstring _achartowstring(const char *s, size_t len) {
     wchar sbuf[512];
     int sz;
 
     if (len == 0)
 	return L"";
-    else if ((sz = MultiByteToWideChar(CP_UTF8, 0, s, len, sbuf,
+    else if ((sz = MultiByteToWideChar(CP_UTF8, 0, s, (int)len, sbuf,
 	sizeof (sbuf) / sizeof (wchar))) > 0)
 	return sbuf;
-    else if ((sz = MultiByteToWideChar(CP_UTF8, 0, s, len, NULL, 0)) == 0)
+    else if ((sz = MultiByteToWideChar(CP_UTF8, 0, s, (int)len, NULL, 0)) == 0)
 	return L"";
 
     wchar *buf = new wchar[(uint)sz];
     wstring ret;
 
-    MultiByteToWideChar(CP_UTF8, 0, s, len, buf, sz);
+    MultiByteToWideChar(CP_UTF8, 0, s, (int)len, buf, sz);
     ret.assign(buf, sz - 1);
     delete [] buf;
     return ret;
 }
 
-const string _wchartoastring(const wchar *s, int len) {
+const string _wchartoastring(const wchar *s, size_t len) {
     char sbuf[1024];
     int sz;
 
     if (len == 0)
 	return "";
-    else if ((sz = WideCharToMultiByte(CP_UTF8, 0, s, len, sbuf, sizeof (sbuf),
-	NULL, NULL)) > 0)
+    else if ((sz = WideCharToMultiByte(CP_UTF8, 0, s, (int)len, sbuf,
+	sizeof (sbuf), NULL, NULL)) > 0)
 	return sbuf;
-    else if ((sz = WideCharToMultiByte(CP_UTF8, 0, s, len, NULL, 0, NULL,
+    else if ((sz = WideCharToMultiByte(CP_UTF8, 0, s, (int)len, NULL, 0, NULL,
 	NULL)) == 0)
 	return "";
 
     char *buf = new char[(uint)sz];
     string ret;
 
-    WideCharToMultiByte(CP_UTF8, 0, s, len, buf, len, NULL, NULL);
+    WideCharToMultiByte(CP_UTF8, 0, s, (int)len, buf, (int)len, NULL, NULL);
     ret.assign(buf, sz - 1);
     delete [] buf;
     return ret;
